Reasoning-Diagnostic-Agent/requirements.txt
a2a-sdk==0.3.10
chromadb==1.1.1
fastmcp==2.12.4
langchain==0.3.27
langchain_ollama==0.3.10
langchain-mcp-adapters==0.1.11
langgraph==0.6.9
langgraph.prebuilt==0.6.4
mcp==1.16.0
sentence-transformers==5.1.1
